# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentary of the Rubik's Market Laravel project",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/admin/blog": {
            "get": {
                "tags": [
                    "Admin Blog"
                ],
                "summary": "List of articles",
                "operationId": "7cc17c3351f422a5085af7455e766de5",
                "responses": {
                    "200": {
                        "description": "Show blog ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/BlogListResource"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin Blog"
                ],
                "summary": "Create article",
                "operationId": "cad647743d9eaa0ebeedc4fefbcedec2",
                "parameters": [
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Title of the article",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "content",
                        "in": "query",
                        "description": "The text of the article",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Photo of the article and categories",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "image",
                                    "categories"
                                ],
                                "properties": {
                                    "image": {
                                        "description": "Photo of the article",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "categories": {
                                        "description": "Category related to the article",
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Create blog"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin/blog/{id}": {
            "get": {
                "tags": [
                    "Admin Blog"
                ],
                "summary": "Get article",
                "operationId": "c59b00450b654281a675cb20cdf78b70",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the blog ",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show blog ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/BlogDetailResource"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Admin Blog"
                ],
                "summary": "Update article",
                "operationId": "ee31d1e9384e66697a1d2fb2e2c9315b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the article",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Title of the article",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "content",
                        "in": "query",
                        "description": "The text of the article",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Photo of the article and categories",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "image",
                                    "categories"
                                ],
                                "properties": {
                                    "image": {
                                        "description": "Photo of the article",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "categories": {
                                        "description": "Category related to the article",
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Update  article "
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin Blog"
                ],
                "summary": "Delete blog",
                "operationId": "931f973c7e99b489c9d5ecdcfd654cd0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the article",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete  blog "
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Admin Blog"
                ],
                "summary": "Update blog",
                "operationId": "238f28fb789213a039fefb3c29fad0b6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the article",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "title",
                        "in": "query",
                        "description": "ID of the blog ",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "discription",
                        "in": "query",
                        "description": "ID of the blog ",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "image",
                        "in": "query",
                        "description": "ID of the blog ",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "binary"
                        }
                    },
                    {
                        "name": "categories",
                        "in": "query",
                        "description": "ID of the blog ",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Craete  blog "
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "BlogDetailResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "بررسی اجمالی آیفون 13: نسل جدید قدرت و نوآوری"
                    },
                    "count_view": {
                        "type": "integer",
                        "example": 25
                    },
                    "content": {
                        "type": "string",
                        "example": "سامسونگ گلکسی S21، یکی از پرچم‌داران قدرتمند سامسونگ، با طراحی زیبا و مشخصات فنی بالا، یکی از بهترین انتخاب‌ها در دنیای گوشی‌های هوشمند به شمار می‌رود. این دستگاه با صفحه‌نمایش **Dynamic AMOLED 2X** و اندازه 6.2 اینچی، تجربه‌ی بصری فوق‌العاده‌ای را برای کاربران فراهم می‌کند. کیفیت رنگ‌ها، کنتراست بالا و نرخ نوسازی 120 هرتز، تجربه کاربری روان و لذت‌بخشی را به همراه دارد.\n\n            چیپست قدرتمند **Exynos 2100** (برای نسخه‌های جهانی) یا **Qualcomm Snapdragon 888** (برای نسخه‌های آمریکا و چین)، سرعت و قدرت فوق‌العاده‌ای برای اجرای برنامه‌ها و بازی‌های سنگین ارائه می‌دهد. با 8 گیگابایت حافظه رم و 128 یا 256 گیگابایت فضای ذخیره‌سازی، گلکسی S21 عملکردی بی‌نقص و سریع دارد. همچنین، این گوشی فاقد درگاه کارت حافظه است، اما فضای ذخیره‌سازی کافی برای اکثر کاربران فراهم می‌کند.\n\n            دوربین‌های سه‌گانه در پشت دستگاه شامل یک لنز 12 مگاپیکسلی فوق عریض، یک لنز 12 مگاپیکسلی عریض، و یک لنز تله‌فوتو 64 مگاپیکسلی است که تجربه‌ای بی‌نظیر از عکاسی و فیلم‌برداری به ارمغان می‌آورد. قابلیت **8K video recording** و **Super Steady** در فیلم‌برداری، باعث می‌شود کاربران بتوانند ویدیوهایی با وضوح بسیار بالا و پایداری بی‌نظیر ضبط کنند.\n\n            **باتری 4000 میلی‌آمپرساعتی** این گوشی با پشتیبانی از شارژ سریع 25 واتی، شارژ بی‌سیم 15 واتی و شارژ بی‌سیم معکوس، امکان استفاده طولانی‌مدت از گوشی بدون نگرانی از تمام شدن باتری را فراهم می‌کند. یکی دیگر از ویژگی‌های برجسته گلکسی S21، پشتیبانی از شبکه 5G است که سرعت بسیار بالایی را در اتصال به اینترنت و دانلود محتوا فراهم می‌کند.\n\n            گلکسی S21 با **استاندارد IP68** در برابر آب و گرد و غبار مقاوم است، به این معنی که در شرایط مختلف محیطی، همچنان عملکرد خود را حفظ می‌کند. سامسونگ همچنین با ارائه تکنولوژی صوتی **Dolby Atmos** و اسپیکرهای استریو، تجربه شنیداری عالی را برای کاربران فراهم کرده است.\n\n            سامسونگ گلکسی S21 از **One UI 3.1**، جدیدترین نسخه رابط کاربری سامسونگ بر پایه اندروید 11 بهره می‌برد که تجربه کاربری ساده، روان و شخصی‌سازی‌شده را به همراه دارد. در مجموع، این گوشی ترکیبی از قدرت، زیبایی و تکنولوژی‌های مدرن است که نیازهای کاربران حرفه‌ای و علاقه‌مندان به دنیای فناوری را به‌خوبی برآورده می‌کند."
                    },
                    "image": {
                        "type": "string",
                        "example": "/storage/files/OIN9JWtZXPg3M9OBnVu7adsaEePPkbN0LVFOOEYgHw.jpg"
                    },
                    "failed_at": {
                        "type": "string",
                        "example": "مهر 10، 1403"
                    }
                },
                "type": "object"
            },
            "BlogListResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "بررسی اجمالی آیفون 13: نسل جدید قدرت و نوآوری"
                    },
                    "count_view": {
                        "type": "integer",
                        "example": 25
                    },
                    "content": {
                        "type": "string",
                        "example": "آیفون ۱۳، محصول جدید اپل، با طراحی زیبا و مشخصات فنی بهبود یافته به بازار عرضه شده است. این دستگاه با صفحه‌نمایش Super Retina XDR، تجربه بصری فوق‌العاده‌ای را به کاربران ارائه می‌دهد. چیپست A۱۵ Bionic، که به عنوان قلب تپنده آیفون ۱۳ عمل می‌کند، عملکرد بی‌نظیری را در اجرای برنامه‌ها و بازی‌های سنگین فراهم می‌سازد. دوربین‌های دوتایی ۱۲ مگاپیکسلی در پشت دستگاه با قابلیت‌های عکاسی پیشرفته مانند حالت شب و Smart HDR ۴، عکس‌ها و ویدیوهایی با کیفیت بالا ثبت می‌کنند."
                    },
                    "image": {
                        "type": "string",
                        "example": "/storage/files/OIN9JWtZXPg3M9OBnVu7adsaEePPkbN0LVFOOEYgHw.jpg"
                    },
                    "failed_at": {
                        "type": "string",
                        "example": "مهر 10، 1403"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}